Clean up code related to json directories.
There are several cleanups here:

    1. json_root is no longer a global variable

    2. We extract settings.py (more on this below).

    3. We check up front to make sure settings.json_directory is
       a valid directory, and fail fast if it isn't (plus
       give helpful suggestions on how to create it).

Windows paths:

I have been testing this using git bash on Windows, which
is hopefully the most challenging test of whether we're
handling windows paths correctly.  You'll see one comment
in this commit related to that.

Settings:

It's very inconvenient to have non-api-key settings inside of zuliprc.
First, we can't provide an out-of-the-box settings file that just
works, because if the zuliprc contains your API key, we
obviously can't supply the API key.  And if you need to update
your API key, you have to download a new zuliprc, which has the
pitfall that you're either gonna clobber your custom settings
related to json_directory/archive_root/etc., or you have to patch
it back in.  Finally, if you want to work with another developer
to troubleshoot your settings, you need to either mask your API
key or trust the other person (particularly troublesome when working
over the web) or just send excerpts of the file (which makes it
harder for the other person to troubleshoot).

And then on top of that, the configparser API is just kinda
clumsy.

I considered having a JSON configuration for all the non-Zulip
stuff, but you can't have comments in JSON.  I mean, you can have
comments in JSON, but then you need to strip them out.

I decided it's easy to just have the settings in Python.  This
is a time-tested technique from Django, except we don't have all
the complexity of a Django settings.py file.

It's particularly nice when a lot of the settings are directory
names to be able to configure them directly in Python.  This isn't
apparent in this commit, since I've only moved one setting so far,
but this will be more obvious later.

Finally, I haven't completely address the source control issues,
but here's the progression:

    BEFORE - no sample configuration files, TOTALLY UNHELPFUL

    NOW - settings.py is under source control, which means
        you can't modify it without updating the repo, but this
        is just temporary

    NEAR FUTURE - I'll move settings.py to something like
        settings.py.template, which will be slightly annoying
        but better than the alternatives.

Sorry this commit message is so long--subsequent commits should
make the strategy more clear here.