Extract libs: populate and common.
I wanted to extract populate.py, since it's
kind of the code that we LEAST want folks to mess
with.  Even though the code's a bit messy, as long
as it moves data into JSON files reliably, nobody
should want to mess with this.

I expect a lot more development to happen related
to building out the HTML/Markdown site, so hopefully
that code will be a bit easier to find now.