Change the default site url to be port 4000.
The default value for site url was a file path,
which was broken for Windows when your source
code was on the `C:\` drive, for typical stupid
Windows-related reasons.  (You ended up getting
'file:///' with three slashes.)

Realistically, more people are likely to test
this locally with Jekyll for now, since we kind
of assume folks are using Jekyll in terms of how
we build the site (using "front matter", for example).

Note that folks have always had the option of
setting this explicitly, and they would have needed
to do that in production.