Fix recent regression in stream/topic links.
6bc94ae3601d54cd121121323b885a8aee71bd46 broke our
stream-related links due to passing 3 parameters into
`urllib.parse.urljoin`.  That API has the nasty
gotcha that the third parameter is an optional
boolean called `allow_fragments`, so the stream
name got ignored (and instead was just intepreted
as a truthy value to allow fragments).

This also cleans up a few redundant calls
to sanitize_stream.