Clean up format_stream_url/format_topic_url.
Before this change `format_stream_url` was returning
this kind of useless intermediate value that was the
prefix to actual stream url, which also includes
`index.html`, and from which we built the topic url
using kind of hard-to-read code.

Now we just have two-liner helpers for both that build
complete urls, so that the calling code can just
do the HTML piece of making the anchor tag or
markdown link.