Extract lib/html.py.
The new html.py does the most important part of creating
an archive, which is actually building out the HTML.  (Ok,
populate.py is pretty important too, but hopefully that's
pretty stable.)

Right now it's only ~250 lines of code.

The next steps here are to make it less Jekyll-centric.

There are some nice things about Jekyll that we'll want to
keep, but we'll want non-Jekyll alternatives for most
stuff:

    - instead of Jekyll front matter, just write
      an ordinary HTML <head> block

    - try to avoid markdown in the final product
      for non-Jekyll cases (even if it's just a matter
      of pumping markdown snippets through a Python
      library)

    - avoid markdown where the HTML is nearly trivial,
      anyway (such as headings)