Remove write_topic_body().
This function had a kind of a misleading name (it would
be more clear to say it writes a list of messages), and
it delegated half of its work in a strange way to
format_message().

Now we just call format_message() directly, and let it
do all the work.