Fix zulip icon url.
We just force folks to set an explicit url, instead
of trying to calculate it relative to the project root.
This gives them flexibility to serve images where they
want, and it should be easier to troubleshoot than
reading the code to figure out how it's generated.

This was broken until now, since our repo didn't have
the asset to begin with (I borrowed it from the Lean
repo), and the url logic was broken due to trailing slash
issues.