Revamp the setup process.
Fix #33, #40, #43.

- entrypoint.sh: Remove the code to automatically enable GH Pages, because `secrets.GITHUB_TOKEN` doesn't have the permission to do so.
- Update README.md to reflect the change above.

I have tried the current workflow with `secrets.GITHUB_TOKEN`. It turns out that now by default the token doesn't have the permission to enable GH Pages. And even after I had increased the privilege to the highest level, it still didn't have enough permission to do so.

In this PR, I change it so that only read-only permission is required, and thus simplifying the setup process. No need to go through the web UI to increase the scope of `secrets.GITHUB_TOKEN`.